Basic GUI for connect four added.
GUI package created for GUI class(es).
GUIConnectFourBoard class for showing pane with board on it.
This class is instantiated in main and added to the main stage. It cfurrently constructs allg odd.
It will have an update method that is called on it to change the GUI depending on the state of the ConnectFourBoard field object and this will be called after the logic is excecuted in the main class